refactor: remove unused farm resource comp and related features, general cleanup
