+++ 1.5/Defs/Buildings/Buildings_Structure_Rustic.xml
@@ -56,19 +56,6 @@
             <li MayRequire="VanillaExpanded.VNutrientE" Class="PipeSystem.CompProperties_Resource">
                 <pipeNet>VNPE_NutrientPasteNet</pipeNet>
             </li>
-            <!-- <li MayRequire="VanillaExpanded.VNutrientE" Class="PawnStorages.VEF.CompProperties_FarmResourceStorage"> -->
-            <!--     <pipeNet>VNPE_NutrientPasteNet</pipeNet> -->
-            <!--     <barSize>(0.75, 0.20)</barSize> -->
-            <!--     <storageCapacity>50</storageCapacity> -->
-            <!--     <margin>0.05</margin> -->
-            <!--     <centerOffset>(0, 0, -0.35)</centerOffset> -->
-            <!--     <destroyOption> -->
-            <!--         <filth>VNPE_Filth_Paste</filth> -->
-            <!--         <ratio>1</ratio> -->
-            <!--         <maxRadius>5</maxRadius> -->
-            <!--     </destroyOption> -->
-            <!--     <contentRequirePower>false</contentRequirePower> -->
-            <!-- </li> -->
         </comps>
         <building>
             <wantsHopperAdjacent>true</wantsHopperAdjacent>
@@ -175,19 +162,6 @@
             <li MayRequire="VanillaExpanded.VNutrientE" Class="PipeSystem.CompProperties_Resource">
                 <pipeNet>VNPE_NutrientPasteNet</pipeNet>
             </li>
-            <!-- <li MayRequire="VanillaExpanded.VNutrientE" Class="PawnStorages.VEF.CompProperties_FarmResourceStorage"> -->
-            <!--     <pipeNet>VNPE_NutrientPasteNet</pipeNet> -->
-            <!--     <barSize>(0.75, 0.20)</barSize> -->
-            <!--     <storageCapacity>50</storageCapacity> -->
-            <!--     <margin>0.05</margin> -->
-            <!--     <centerOffset>(0, 0, -0.35)</centerOffset> -->
-            <!--     <destroyOption> -->
-            <!--         <filth>VNPE_Filth_Paste</filth> -->
-            <!--         <ratio>1</ratio> -->
-            <!--         <maxRadius>5</maxRadius> -->
-            <!--     </destroyOption> -->
-            <!--     <contentRequirePower>false</contentRequirePower> -->
-            <!-- </li> -->
         </comps>
         <building>
             <wantsHopperAdjacent>true</wantsHopperAdjacent>
