--- 1.5/Source/PawnStorages/PawnStoragesModule-VEF/CompProperties_FarmResourceStorage.cs
@@ -1,9 +1,0 @@
-﻿using PipeSystem;
-
-namespace PawnStorages.VEF;
-
-public class CompProperties_FarmResourceStorage : CompProperties_NutritionNetStorage
-{
-    public CompProperties_FarmResourceStorage() => this.compClass = typeof (CompFarmResourceStorage);
-
-}
