--- 1.5/Source/PawnStorages/PawnStoragesModule-VEF/CompProperties_NutritionNetStorage.cs
@@ -1,8 +1,0 @@
-﻿using PipeSystem;
-
-namespace PawnStorages.VEF;
-
-public class CompProperties_NutritionNetStorage : CompProperties_ResourceStorage
-{
-    public CompProperties_NutritionNetStorage() => compClass = typeof (CompNutritionNetStorage);
-}
