--- 1.5/Source/PawnStorages/PawnStoragesModule-VEF/HarmonyPatches/Building_PSFarm_Patches.cs
@@ -1,10 +1,0 @@
-﻿using HarmonyLib;
-using PawnStorages.Farm;
-
-namespace PawnStorages.VEF.HarmonyPatches;
-
-[HarmonyPatch(typeof(Building_PSFarm))]
-public static class Building_PSFarm_Patches
-{
-
-}
